pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel", "torch>=2.8"]
build-backend = "setuptools.build_meta"

[project]
name = "gYCSB"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "PyYAML==6.0.2",
    "dash==3.2.0",
    "dash-bootstrap-components==2.0.4",
    "plotly==6.1.0",
    "pandas",
    "matplotlib",
    "psutil==7.0.0",
    "pyarrow==20.0.0",
    "torch>=2.8",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "flask",
    "flask-cors"
]

[tool.setuptools.packages.find]
where = ["."]
